throw custom error

+++ src/storage/group/groupCreate.ts
@@ -1,10 +1,18 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { GROUP_COLLECTION } from '@storage/sotrageConfig';
 import { getGroups } from './getGroups';
+
+import { AppError } from '@utils/AppError';
+import { GROUP_COLLECTION } from '@storage/sotrageConfig';
 
 export async function groupCreate(newGroup: string){
   try {
     const storedGroups = await getGroups();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if(groupAlreadyExists){
+      throw new AppError('Já existe um grupo cadastrado com esse nome')
+    }
 
     const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
